refactor: Extract Datacards component
